Bug #1578919 bsp: camera: Fix converity warning
[root cause  ] value type not compatible
[changes     ] isp_drv.c
[side effects] NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums9620/ums9230
[test case]:camera function
[reviewers   ] yangyang.liu
[change_type ] feature_bugfix
[tag_product ] common

Change the value type to fix converity warning.

+++ core/isp2.6/src/isp_drv.c
@@ -156,7 +156,7 @@ static int ispdrv_fetch_normal_get(void *cfg_in, void *cfg_out,
 		struct camera_frame *frame)
 {
 	int ret = 0;
-	unsigned long trim_offset[3] = { 0 };
+	uint32_t trim_offset[3] = { 0 };
 	struct img_size *src = NULL;
 	struct img_trim *intrim = NULL;
 	struct isp_hw_fetch_info *fetch = NULL;
